+++ src/background.js
@@ -20,7 +20,8 @@ const NXTK = (() => {
     NDC_pauseBetweenDownload: 5,
     NDC_downloadSpeed: 3.2,
     ForceEnglish: false,
-    NDC_downloadMethod: 0
+    NDC_downloadMethod: 0,
+    WabbajackImport: false
   };
 
   const SENSITIVE_PARAM_NAMES = [